Endret noen val navn og flyttet data klasse PeriodeSoekRequest i egen fil
